Move alertFail to utils
